requirements.txt
fastapi
uvicorn
psycopg2-binary
pytest
pytest-asyncio